fix: selecting moves for fight commando does not show not found message anymore
